Refactor list_box_model example
