+++ examples/list_box_model/README.md
@@ -1,7 +1,8 @@
 # ListBox and ListModel example
 
 This example demonstrates how to use `gtk::ListBox` in combination with
-`gio::ListStore` as a model with a custom row type.
+a custom model and a custom row type.
+The custom model resides in `model.rs` and the custom row type in `row_data.rs`.
 
 It sets up a `gtk::ListBox` containing, per row, a label, spinbutton and
 an edit button. The edit button allows to edit the underlying data structure
