+++ examples/list_box_model/main.rs
@@ -1,88 +1,13 @@
+mod model;
+pub mod row_data;
+
 use gtk::{
-    gio,
     glib::{self, clone},
     prelude::*,
     ResponseType,
 };
-
-use std::env::args;
-
 use row_data::RowData;
-
-mod model {
-    use super::*;
-    use glib::subclass::prelude::*;
-    use row_data::RowData;
-
-    mod imp {
-        use super::*;
-        use gio::subclass::prelude::*;
-        use std::cell::RefCell;
-
-        #[derive(Debug, Default)]
-        pub struct Model(pub RefCell<Vec<RowData>>);
-
-        // Basic declaration of our type for the GObject type system
-        #[glib::object_subclass]
-        impl ObjectSubclass for Model {
-            const NAME: &'static str = "Model";
-            type Type = super::Model;
-            type ParentType = glib::Object;
-            type Interfaces = (gio::ListModel,);
-        }
-
-        impl ObjectImpl for Model {}
-
-        impl ListModelImpl for Model {
-            fn get_item_type(&self, _list_model: &Self::Type) -> glib::Type {
-                RowData::static_type()
-            }
-            fn get_n_items(&self, _list_model: &Self::Type) -> u32 {
-                self.0.borrow().len() as u32
-            }
-            fn get_item(&self, _list_model: &Self::Type, position: u32) -> Option<glib::Object> {
-                self.0
-                    .borrow()
-                    .get(position as usize)
-                    .map(|o| o.clone().upcast::<glib::Object>())
-            }
-        }
-    }
-
-    // Public part of the Model type.
-    glib::wrapper! {
-        pub struct Model(ObjectSubclass<imp::Model>) @implements gio::ListModel;
-    }
-
-    // Constructor for new instances. This simply calls glib::Object::new()
-    impl Model {
-        #[allow(clippy::new_without_default)]
-        pub fn new() -> Model {
-            glib::Object::new(&[]).expect("Failed to create Model")
-        }
-
-        pub fn append(&self, obj: &RowData) {
-            let self_ = imp::Model::from_instance(self);
-            let index = {
-                // Borrow the data only once and ensure the borrow guard is dropped
-                // before we emit the items_changed signal because the view
-                // could call get_item / get_n_item from the signal handler to update its state
-                let mut data = self_.0.borrow_mut();
-                data.push(obj.clone());
-                data.len() - 1
-            };
-            // Emits a signal that 1 item was added, 0 removed at the position index
-            self.items_changed(index as u32, 0, 1);
-        }
-
-        pub fn remove(&self, index: u32) {
-            let self_ = imp::Model::from_instance(self);
-            self_.0.borrow_mut().remove(index as usize);
-            // Emits a signal that 1 item was removed, 0 added at the position index
-            self.items_changed(index, 1, 0);
-        }
-    }
-}
+use std::env::args;
 
 fn build_ui(application: &gtk::Application) {
     let window = gtk::ApplicationWindow::new(application);
@@ -261,123 +186,4 @@ fn main() {
     });
 
     application.run(&args().collect::<Vec<_>>());
-}
-
-// Our GObject subclass for carrying a name and count for the ListBox model
-//
-// Both name and count are stored in a RefCell to allow for interior mutability
-// and are exposed via normal GObject properties. This allows us to use property
-// bindings below to bind the values with what widgets display in the UI
-mod row_data {
-    use super::*;
-
-    use glib::subclass::prelude::*;
-
-    // Implementation sub-module of the GObject
-    mod imp {
-        use super::*;
-        use std::cell::RefCell;
-
-        // The actual data structure that stores our values. This is not accessible
-        // directly from the outside.
-        #[derive(Default)]
-        pub struct RowData {
-            name: RefCell<Option<String>>,
-            count: RefCell<u32>,
-        }
-
-        // Basic declaration of our type for the GObject type system
-        #[glib::object_subclass]
-        impl ObjectSubclass for RowData {
-            const NAME: &'static str = "RowData";
-            type Type = super::RowData;
-            type ParentType = glib::Object;
-        }
-
-        // The ObjectImpl trait provides the setters/getters for GObject properties.
-        // Here we need to provide the values that are internally stored back to the
-        // caller, or store whatever new value the caller is providing.
-        //
-        // This maps between the GObject properties and our internal storage of the
-        // corresponding values of the properties.
-        impl ObjectImpl for RowData {
-            fn properties() -> &'static [glib::ParamSpec] {
-                use once_cell::sync::Lazy;
-                static PROPERTIES: Lazy<Vec<glib::ParamSpec>> = Lazy::new(|| {
-                    vec![
-                        glib::ParamSpec::string(
-                            "name",
-                            "Name",
-                            "Name",
-                            None, // Default value
-                            glib::ParamFlags::READWRITE,
-                        ),
-                        glib::ParamSpec::uint(
-                            "count",
-                            "Count",
-                            "Count",
-                            0,
-                            100,
-                            0, // Allowed range and default value
-                            glib::ParamFlags::READWRITE,
-                        ),
-                    ]
-                });
-
-                PROPERTIES.as_ref()
-            }
-
-            fn set_property(
-                &self,
-                _obj: &Self::Type,
-                _id: usize,
-                value: &glib::Value,
-                pspec: &glib::ParamSpec,
-            ) {
-                match pspec.get_name() {
-                    "name" => {
-                        let name = value
-                            .get()
-                            .expect("type conformity checked by `Object::set_property`");
-                        self.name.replace(name);
-                    }
-                    "count" => {
-                        let count = value
-                            .get_some()
-                            .expect("type conformity checked by `Object::set_property`");
-                        self.count.replace(count);
-                    }
-                    _ => unimplemented!(),
-                }
-            }
-
-            fn get_property(
-                &self,
-                _obj: &Self::Type,
-                _id: usize,
-                pspec: &glib::ParamSpec,
-            ) -> glib::Value {
-                match pspec.get_name() {
-                    "name" => self.name.borrow().to_value(),
-                    "count" => self.count.borrow().to_value(),
-                    _ => unimplemented!(),
-                }
-            }
-        }
-    }
-
-    // Public part of the RowData type. This behaves like a normal gtk-rs-style GObject
-    // binding
-    glib::wrapper! {
-        pub struct RowData(ObjectSubclass<imp::RowData>);
-    }
-
-    // Constructor for new instances. This simply calls glib::Object::new() with
-    // initial values for our two properties and then returns the new instance
-    impl RowData {
-        pub fn new(name: &str, count: u32) -> RowData {
-            glib::Object::new(&[("name", &name), ("count", &count)])
-                .expect("Failed to create row data")
-        }
-    }
 }
