+++ examples/list_box_model/row_data.rs
@@ -1,0 +1,115 @@
+//! Our GObject subclass for carrying a name and count for the ListBox model
+//!
+//! Both name and count are stored in a RefCell to allow for interior mutability
+//! and are exposed via normal GObject properties. This allows us to use property
+//! bindings below to bind the values with what widgets display in the UI
+
+use glib::subclass::prelude::*;
+use gtk::{glib, prelude::*};
+
+// Implementation sub-module of the GObject
+mod imp {
+    use super::*;
+    use std::cell::RefCell;
+
+    // The actual data structure that stores our values. This is not accessible
+    // directly from the outside.
+    #[derive(Default)]
+    pub struct RowData {
+        name: RefCell<Option<String>>,
+        count: RefCell<u32>,
+    }
+
+    // Basic declaration of our type for the GObject type system
+    #[glib::object_subclass]
+    impl ObjectSubclass for RowData {
+        const NAME: &'static str = "RowData";
+        type Type = super::RowData;
+        type ParentType = glib::Object;
+    }
+
+    // The ObjectImpl trait provides the setters/getters for GObject properties.
+    // Here we need to provide the values that are internally stored back to the
+    // caller, or store whatever new value the caller is providing.
+    //
+    // This maps between the GObject properties and our internal storage of the
+    // corresponding values of the properties.
+    impl ObjectImpl for RowData {
+        fn properties() -> &'static [glib::ParamSpec] {
+            use once_cell::sync::Lazy;
+            static PROPERTIES: Lazy<Vec<glib::ParamSpec>> = Lazy::new(|| {
+                vec![
+                    glib::ParamSpec::string(
+                        "name",
+                        "Name",
+                        "Name",
+                        None, // Default value
+                        glib::ParamFlags::READWRITE,
+                    ),
+                    glib::ParamSpec::uint(
+                        "count",
+                        "Count",
+                        "Count",
+                        0,
+                        100,
+                        0, // Allowed range and default value
+                        glib::ParamFlags::READWRITE,
+                    ),
+                ]
+            });
+
+            PROPERTIES.as_ref()
+        }
+
+        fn set_property(
+            &self,
+            _obj: &Self::Type,
+            _id: usize,
+            value: &glib::Value,
+            pspec: &glib::ParamSpec,
+        ) {
+            match pspec.get_name() {
+                "name" => {
+                    let name = value
+                        .get()
+                        .expect("type conformity checked by `Object::set_property`");
+                    self.name.replace(name);
+                }
+                "count" => {
+                    let count = value
+                        .get_some()
+                        .expect("type conformity checked by `Object::set_property`");
+                    self.count.replace(count);
+                }
+                _ => unimplemented!(),
+            }
+        }
+
+        fn get_property(
+            &self,
+            _obj: &Self::Type,
+            _id: usize,
+            pspec: &glib::ParamSpec,
+        ) -> glib::Value {
+            match pspec.get_name() {
+                "name" => self.name.borrow().to_value(),
+                "count" => self.count.borrow().to_value(),
+                _ => unimplemented!(),
+            }
+        }
+    }
+}
+
+// Public part of the RowData type. This behaves like a normal gtk-rs-style GObject
+// binding
+glib::wrapper! {
+    pub struct RowData(ObjectSubclass<imp::RowData>);
+}
+
+// Constructor for new instances. This simply calls glib::Object::new() with
+// initial values for our two properties and then returns the new instance
+impl RowData {
+    pub fn new(name: &str, count: u32) -> RowData {
+        glib::Object::new(&[("name", &name), ("count", &count)]).expect("Failed to create row data")
+    }
+}
